fixed problems, after review
