--- src/App.css
@@ -1,18 +1,0 @@
-.container{
-    width: 1240px;
-    margin: 0 auto;
-}
-a{
-    color: inherit;
-    text-decoration: none;
-}
-.burgerSectionContainer{
-    display: flex;
-    justify-content: space-between;
-}
-.burger-constructor-container .constructor-element{
-    flex-grow: 0 !important;
-    max-width: 536px;
-    min-width: 536px;
-    width: 100%;
-}
