+++ src/components/app-header/app-header.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import {BurgerIcon, ListIcon, Logo, ProfileIcon} from "@ya.praktikum/react-developer-burger-ui-components";
 import styles from './app-header.module.css'
-function AppHeader(props: any) {
+//@ts-ignore
+function AppHeader() {
     return (
         <header className={`pb-4 pt-4 ${styles.mainHeader}`}>
             <div className="container">
