+++ src/components/app/app.module.css
@@ -1,0 +1,4 @@
+.burgerSectionContainer{
+    display: flex;
+    justify-content: space-between;
+}
