+++ src/index.css
@@ -14,3 +14,17 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+.container{
+  width: 1240px;
+  margin: 0 auto;
+}
+a{
+  color: inherit;
+  text-decoration: none;
+}
+.burger-constructor-container .constructor-element{
+  flex-grow: 0 !important;
+  max-width: 536px;
+  min-width: 536px;
+  width: 100%;
+}
